beeld/home/serve/requirements.txt
tensorflow-cpu
fastapi
Pillow
starlette==0.22.0
gunicorn
uvicorn
numpy
python-multipart